requirements.txt
gpio==0.3.0
paho-mqtt==1.5.1
py532lib==2018.6.14.dev1
